pyproject.toml
[tool.poetry]
name = "opencrags-rest-api"
version = "0.0.0"
description = ""
authors = ["Richard Löwenström"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.65.1"
uvicorn = "^0.13.4"
requests = "^2.25.1"
Pillow = "^8.2.0"
fastapi-auth0 = "^0.2.1"
pymongo = "^3.11.4"
odmantic = "^0.3.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
